test(agentic-ai): split OpenAI v2 subprocess test base by API family, drop server-tool test
Introduce BaseOpenAiResponsesSubProcessTest as the Responses sibling of
BaseOpenAiCompletionsSubProcessTest (renamed from BaseOpenAiSubProcessTest),
each owning only its own family's provider/backend wiring - mirroring the
existing Anthropic base. AgentSubProcessOpenAiResponsesAdvancedFeaturesTests
now extends the new Responses base instead of BaseAgentSubProcessV2Test
directly, dropping its inline provider-configurer method.

Also removes the server-tool round-trip test
(unmappedServerToolItemRoundTripsVerbatimAsProviderContentOnFollowUpRequest)
and its supporting ServerToolTurnStub/stubServerToolConversation stub code:
the native OpenAI provider does not support server tools yet.

Trims the remaining javadocs to describe current behavior rather than the
design process that produced it, dropping "design spec"/"witness N"
terminology and implementation-history references in favor of concise,
reader-facing descriptions.
